Refactor mongoose.js file
